+++ src/main/java/io/github/opencubicchunks/cubicchunks/mixin/core/client/renderer/chunk/MixinRenderRegionCache.java
@@ -1,4 +1,4 @@
-package io.github.opencubicchunks.cubicchunks.mixin.core.common.client.renderer.chunk;
+package io.github.opencubicchunks.cubicchunks.mixin.core.client.renderer.chunk;
 
 import javax.annotation.Nullable;
