+++ src/main/java/io/github/opencubicchunks/cubicchunks/mixin/core/client/renderer/cube/MixinRenderCube.java
@@ -1,4 +1,4 @@
-package io.github.opencubicchunks.cubicchunks.mixin.core.common.client.renderer.cube;
+package io.github.opencubicchunks.cubicchunks.mixin.core.client.renderer.cube;
 
 import io.github.opencubicchunks.cc_core.utils.Coords;
 import io.github.opencubicchunks.cubicchunks.client.renderer.cube.RenderCube;
